[add-single-product] chore : 변수형 int -> BigInteger로 변경

+++ src/main/java/hello/fclover/domain/GoodsImage.java
@@ -5,13 +5,15 @@ import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
 
+import java.math.BigInteger;
+
 @Getter
 @Setter
 @ToString
 @Data
 public class GoodsImage {
-    private int imageNo; // 이미지 키
-    private int goodsNo; // 상품 번호
+    private BigInteger imageNo; // 이미지 키
+    private BigInteger goodsNo; // 상품 번호
     private String goodsImageName; // 이미지 이름
     private String goodsUrl; //이미지 저장 path
     private String isMain; //대표이미지 설정
